requirements-simple.txt
bottle==0.12.10
lxml==3.6.4
requests==2.12.3
wit==4.2.0
python-Levenshtein==0.12.0